refactor(api): refactored file structure to confrom to best practices
